requirements.txt
numpy>=1.24
pandas>=2.0
xarray>=2024.1
netCDF4>=1.6
scipy>=1.11
torch>=2.1
torch-geometric>=2.4
matplotlib>=3.7
ffmpeg-python>=0.2
google-genai>=0.4
chromadb>=0.4
Pillow>=10.0
streamlit>=1.30
tqdm>=4.65
scikit-learn>=1.3
